+++ src/components/enum/module.ts
@@ -4,12 +4,17 @@ import {FormsModule} from '@angular/forms';
 
 import {CsIconModule} from '../icon/module';
 
+import {CsEnum} from './enum-meta';
+import {CsEnumPipe} from './enum-pipe';
 import {CsEnumSelect} from './enum-select';
 
-export {CsEnumSelect};
+export {CsEnum, CsEnumSelect, CsEnumPipe};
 
 @NgModule({
     imports: [CommonModule, FormsModule, CsIconModule],
-    declarations: [CsEnumSelect]
+    declarations: [
+        CsEnumSelect,
+        CsEnumPipe
+    ]
 })
 export class CsEnumModule { }